refactor - modify order table groups package structure
- 단체석과 관련된 패키지 구조를 에그리거트 중심 구조로 변경했습니다
