+++ src/main/java/kitchenpos/ordertablegroups/bo/TableGroupBo.java
@@ -1,11 +1,11 @@
-package kitchenpos.bo;
+package kitchenpos.ordertablegroups.bo;
 
 import kitchenpos.orders.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
-import kitchenpos.dao.TableGroupDao;
+import kitchenpos.ordertablegroups.dao.TableGroupDao;
 import kitchenpos.orders.model.OrderStatus;
 import kitchenpos.model.OrderTable;
-import kitchenpos.model.TableGroup;
+import kitchenpos.ordertablegroups.model.TableGroup;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
